\name{polyCov-package}
\alias{polyCov-package}
\docType{package}
\title{
Estimate polychoric correlations and covariances.
}
\description{
This is a very small package with only one function, polyCov(). This 
function can be used to estimate a polychoric (or tetrachoric) 
correlation matrix assuming standardized latent items. If 
thresholds are given, it estimates polychoric covariances, means 
and variances.
}
\details{
\tabular{ll}{
Package: \tab polyCov\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-30\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Sacha Epskamp <sacha.epskamp@gmail.com>
}

